Rebase Refactor passes module & clean-up exports with v0.10.0-rc  (#1432)
**Context:** Needed to rebase refactoring branch to test.

**Description of the Change:**
* Adds documentation to Pass, PassPlugin
* Uses values from Pass's options as options in the
ApplyRegisteredPassOps

**Benefits:** Documentation, passes can have options.

---------
