+++ frontend/catalyst/__init__.py
@@ -81,7 +81,7 @@ sys.modules["mlir_quantum._mlir_libs._quantumDialects.mitigation"] = types.Modul
     "mlir_quantum._mlir_libs._quantumDialects.mitigation"
 )
 
-from catalyst import debug, logging
+from catalyst import debug, logging, passes
 from catalyst.api_extensions import *
 from catalyst.api_extensions import __all__ as _api_extension_list
 from catalyst.autograph import *
@@ -89,7 +89,7 @@ from catalyst.autograph import __all__ as _autograph_functions
 from catalyst.compiler import CompileOptions
 from catalyst.debug.assertion import debug_assert
 from catalyst.jit import QJIT, qjit
-from catalyst.passes import Pass, PassPlugin, apply_pass, apply_pass_plugin, pipeline
+from catalyst.passes.pass_api import pipeline
 from catalyst.utils.exceptions import (
     AutoGraphError,
     CompileError,
@@ -187,11 +187,8 @@ __all__ = (
     "debug_assert",
     "CompileOptions",
     "debug",
-    "apply_pass",
-    "apply_pass_plugin",
+    "passes",
     "pipeline",
-    "Pass",
-    "PassPlugin",
     *_api_extension_list,
     *_autograph_functions,
 )
